+++ io/plugins/eu.esdihumboldt.hale.io.gml.ui/src/eu/esdihumboldt/hale/io/gml/ui/PartitionConfigurationPage.java
@@ -53,6 +53,7 @@ public class PartitionConfigurationPage
 	private Button activatePartitioning;
 	private Spinner instances;
 	private ComboViewer partitionMode;
+	private Button activatePartitioningByFeatureType;
 
 	/**
 	 * Default constructor.
@@ -76,11 +77,18 @@ public class PartitionConfigurationPage
 
 	@Override
 	public boolean updateConfiguration(StreamGmlWriter provider) {
-		int threshold = activatePartitioning.getSelection() ? instances.getSelection()
-				: StreamGmlWriter.NO_PARTITIONING;
-		provider.setParameter(StreamGmlWriter.PARAM_INSTANCES_THRESHOLD, Value.of(threshold));
+		if (activatePartitioning.getSelection()) {
+			provider.setParameter(StreamGmlWriter.PARAM_PARTITION_BY_FEATURE_TYPE, Value.of(false));
 
-		applyPartitionMode(partitionMode, provider);
+			int threshold = instances.getSelection();
+			provider.setParameter(StreamGmlWriter.PARAM_INSTANCES_THRESHOLD, Value.of(threshold));
+			applyPartitionMode(partitionMode, provider);
+		}
+		else if (activatePartitioningByFeatureType.getSelection()) {
+			provider.setParameter(StreamGmlWriter.PARAM_INSTANCES_THRESHOLD,
+					Value.of(StreamGmlWriter.NO_PARTITIONING));
+			provider.setParameter(StreamGmlWriter.PARAM_PARTITION_BY_FEATURE_TYPE, Value.of(true));
+		}
 
 		return true;
 	}
@@ -132,6 +140,25 @@ public class PartitionConfigurationPage
 		partitionMode = createPartitionModeSelector(part);
 		GridDataFactory.swtDefaults().align(SWT.FILL, SWT.CENTER)
 				.applyTo(partitionMode.getControl());
+
+		Group gml = new Group(page, SWT.NONE);
+		gml.setLayout(new GridLayout(3, false));
+		gml.setText("Split by feature type");
+		groupData.applyTo(gml);
+
+		activatePartitioningByFeatureType = new Button(gml, SWT.CHECK);
+		activatePartitioningByFeatureType
+				.setText("Create separate output file for every feature type");
+		GridDataFactory.swtDefaults().align(SWT.FILL, SWT.BEGINNING).span(3, 1).grab(true, false)
+				.applyTo(activatePartitioningByFeatureType);
+		activatePartitioningByFeatureType.setSelection(false);
+		activatePartitioningByFeatureType.addSelectionListener(new SelectionAdapter() {
+
+			@Override
+			public void widgetSelected(SelectionEvent e) {
+				update();
+			}
+		});
 
 		update();
 		setPageComplete(true);
@@ -206,6 +233,8 @@ public class PartitionConfigurationPage
 	private void update() {
 		instances.setEnabled(activatePartitioning.getSelection());
 		partitionMode.getControl().setEnabled(activatePartitioning.getSelection());
+		activatePartitioningByFeatureType.setEnabled(!activatePartitioning.getSelection());
+		activatePartitioning.setEnabled(!activatePartitioningByFeatureType.getSelection());
 	}
 
 }